Updating the feature to meet PPAM requirements

+++ app/src/main/java/com/dicoding/ebin_v1/data/response/TrashStationsResponse.kt
@@ -40,7 +40,7 @@ data class TrashStationsResponseItem(
 	val id: String? = null,
 
 	@field:SerializedName("capacity")
-	val capacity: Int? = null
+	val capacity: Float? = null
 ) : Parcelable
 
 @Parcelize
